アプリ側のプロジェを警告をエラーとして扱う設定に変更。UI側の警告をすべて消した。 それに伴ってSpdsのDrawRotaGraphの座標をintからfloatに変更。 メニューを開いた時に一時停止のフラグをオンオフするようにした。

+++ SrcEngine/Drawer/TextDrawer.cpp
@@ -303,7 +303,7 @@ void TextDrawer::DrawString(string str, int32_t x, int32_t y, Align alignment, C
 		break;
 	}
 
-	SpDS::DrawRotaGraph((int32_t)spr->pos.x, (int32_t)spr->pos.y,
+	SpDS::DrawRotaGraph(spr->pos.x, spr->pos.y,
 		spr->scale.x, spr->scale.y, 0.f, spr->key, Anchor::Center, color);
 
 	GetInstance()->releaseQueue.push_back(strData.key);
